Step 1: Get ready
Nothing important, just added a `text` parameter to
`hoedown_document_render` so we get the preprocessed source too,
and modified `char_link` to use `popbuf` instead of that ugly hack.

We also fix `parse_blockquote` to operate in a copy of the input data,
instead of modifying it in place.

Everything should behave the same as before.