Step 2: The mapping stack
Things start to get interesting. We first define a `mapped_chunk` struct. This struct
associates a chunk of data (defined by `buffer`, `offset` and `size`) to the position
of the preprocessed source where that chunk appears.

Thus, we add a stack of `mapped_chunk` objects to `hoedown_document` which will hold
the list of associations. The `map_chunk` method will add an entry to the top of the stack,
with the passed values.