Step 3: Implement source mapping!!
Here's where the magic happens.

First of all, in `hoedown_document_render` we call `map_chunk` to map the preprocessed
source to position 0. After the render finishes, we remove that mapping.

But just one mapping isn't enough, because sometimes intermediate buffers are created
and content is copied to them. We need to keep *those* buffers mapped too. So we create
`custom_buffer_put` which is like `hoedown_buffer_put` but maps the chunk that is being put.

And how do we remove these mappings once the intermediate buffer is released?
By modifying `popbuf`, that's how. Slow, but doesn't need any extra code.

Finally, we export a `hoedown_document_locate` function accepting a pointer to a uint8_t
character. If that character belongs to a chunk that is mapped to a position, we add that
position to the offset of this character in the chunk, and that's it! We have the location
of that character in the source! YES!!1