Step 5: Use custom_buffer_put
In order to highlight all the spans, as we said in step 3, we need to
keep intermediate buffers mapped as well.

Doing that is simple, we just Find & Replace `hoedown_buffer_put`
with `custom_buffer_put` where necessary, and that's all.

As a last note, remember this is *direct* source mapping, it can
only handle data that was copied directly from the source. Using
`custom_buffer_put` on anything else will achieve nothing.